Add parallel specs for categories and users [broken]
I don't know why these are broken. Is it because we're using
fixtures? I don't know.

Pretty much all of the specs that are broken for
models/users are also broken in models/categories,
so I expect this is an issue with the testing environment
rather than an issue with the actual code.