Fix counter_cache to update after move
The children count does not update unless you are creating/deleting
items. For any kind of UI where you are dragging/dropping items around
the current implementation will not update anything.

This commit adds a method to the `Move` callback that updates all items
that are affected by the move.